@@ -3,7 +3,7 @@
   "author": "adaept",
   "homepage": "http://adaept.com/",
   "private": true,
-  "version": "0.2.7",
+  "version": "0.2.8",
   "scripts": {
     "clean": "ionic-app-scripts clean",
     "build": "ionic-app-scripts build",
@@ -11,43 +11,44 @@
     "ionic:serve": "ionic-app-scripts serve"
   },
   "dependencies": {
-    "@angular/common": "4.4.3",
-    "@angular/compiler": "4.4.3",
-    "@angular/compiler-cli": "4.4.3",
-    "@angular/core": "4.4.3",
-    "@angular/forms": "4.4.3",
-    "@angular/http": "4.4.3",
-    "@angular/platform-browser": "4.4.3",
-    "@angular/platform-browser-dynamic": "4.4.3",
-    "@ionic-native/camera": "4.0.0",
-    "@ionic-native/core": "4.0.0",
-    "@ionic-native/google-maps": "4.0.0",
-    "@ionic-native/splash-screen": "4.0.0",
-    "@ionic-native/status-bar": "4.0.0",
+    "@angular/common": "5.0.1",
+    "@angular/compiler": "5.0.1",
+    "@angular/compiler-cli": "5.0.1",
+    "@angular/core": "5.0.1",
+    "@angular/forms": "5.0.1",
+    "@angular/http": "5.0.1",
+    "@angular/platform-browser": "5.0.1",
+    "@angular/platform-browser-dynamic": "5.0.1",
+    "@angular/tsc-wrapped": "5.0.0-beta.7",
+    "@ionic-native/camera": "4.4.2",
+    "@ionic-native/core": "4.4.2",
+    "@ionic-native/google-maps": "4.4.2",
+    "@ionic-native/splash-screen": "4.4.2",
+    "@ionic-native/status-bar": "4.4.2",
     "@ionic/app-scripts": "2.1.3",
     "@ionic/cloud-angular": "^0.12.0",
-    "@ionic/storage": "2.0.1",
-    "@ngx-translate/core": "6.0.1",
-    "@ngx-translate/http-loader": "0.0.3",
+    "@ionic/storage": "2.1.3",
+    "@ngx-translate/core": "8.0.0",
+    "@ngx-translate/http-loader": "2.0.0",
     "@types/greensock": "^1.15.32",
     "a": "^2.1.0",
     "ajv": "^5.2.2",
-    "angularfire2": "5.0.0-rc.0-next",
+    "angularfire2": "5.0.0-rc.4",
     "bible-passage-reference-parser": "^2.0.1",
-    "firebase": "4.4.0",
+    "firebase": "4.6.2",
     "gsap": "^1.20.2",
-    "ionic-angular": "3.7.1",
+    "ionic-angular": "3.9.2",
     "ionicons": "3.0.0",
-    "rxjs": "5.4.3",
-    "sw-toolbox": "3.4.0",
-    "typescript": "2.2.1",
+    "rxjs": "5.5.2",
+    "sw-toolbox": "3.6.0",
+    "typescript": "2.4.2",
     "zone.js": "0.8.18"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "4.4.4",
-    "@ionic/app-scripts": "3.0.0",
+    "@angular/compiler-cli": "5.0.1",
+    "@ionic/app-scripts": "3.1.2",
     "@types/node": "6.0.88",
-    "typescript": "2.2.1"
+    "typescript": "2.4.2"
   },
   "cordovaPlugins": [
     "cordova-plugin-whitelist",
